pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "pop-ml-simulator"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]
